+++ dataland-qa-service/src/test/kotlin/org/dataland/datalandqaservice/services/QaReviewQueryServiceTest.kt
@@ -9,7 +9,6 @@ import org.dataland.datalandbackendutils.model.QaStatus
 import org.dataland.datalandqaservice.org.dataland.datalandqaservice.entities.DatasetJudgementEntity
 import org.dataland.datalandqaservice.org.dataland.datalandqaservice.entities.QaReviewEntity
 import org.dataland.datalandqaservice.org.dataland.datalandqaservice.model.DatasetJudgementState
-import org.dataland.datalandqaservice.org.dataland.datalandqaservice.model.QaReviewResponse
 import org.dataland.datalandqaservice.org.dataland.datalandqaservice.repositories.DatasetJudgementRepository
 import org.dataland.datalandqaservice.org.dataland.datalandqaservice.services.DataPointQaReportManager
 import org.dataland.datalandqaservice.org.dataland.datalandqaservice.services.DatalandBackendAccessor
@@ -83,87 +82,7 @@ class QaReviewQueryServiceTest {
     private val judgementId2: UUID = UUID.randomUUID()
     private val judgeId1: UUID = UUID.randomUUID()
     private val judgeId2: UUID = UUID.randomUUID()
-
-    data class TestSetResult(
-        val qaReview: QaReviewEntity,
-        val datasetJudgement: DatasetJudgementEntity,
-        val response: QaReviewResponse,
-        val dataSourcingPriority: DataSourcingPriorityByDataDimensions,
-    )
-
-    data class BuildTestSetArgs(
-        val dataId: String,
-        val companyId: String,
-        val companyName: String,
-        val reportingPeriod: String,
-        val timestamp: Long = 0L,
-        val triggeringUserId: String = "",
-        val comment: String? = null,
-        val judgementId: UUID,
-        val qaJudgeUserId: UUID,
-        val qaJudgeUserName: String,
-        val numberQaReports: Long = 0L,
-        val priority: Int = 0,
-    )
-
-    private fun buildTestSet(args: BuildTestSetArgs): TestSetResult {
-        val framework = "sfdr"
-        val qaReview =
-            QaReviewEntity(
-                dataId = args.dataId,
-                companyId = args.companyId,
-                companyName = args.companyName,
-                framework = framework,
-                reportingPeriod = args.reportingPeriod,
-                timestamp = args.timestamp,
-                qaStatus = QaStatus.Pending,
-                triggeringUserId = args.triggeringUserId,
-                comment = args.comment,
-            )
-
-        val datasetJudgement =
-            DatasetJudgementEntity(
-                dataSetJudgementId = args.judgementId,
-                datasetId = UUID.fromString(args.dataId),
-                companyId = UUID.randomUUID(),
-                dataType = DataTypeEnum.sfdr,
-                reportingPeriod = args.reportingPeriod,
-                judgementState = DatasetJudgementState.Pending,
-                qaJudgeUserId = args.qaJudgeUserId,
-                qaJudgeUserName = args.qaJudgeUserName,
-                qaReporters = mutableListOf(),
-                dataPoints = mutableListOf(),
-            )
-
-        val response =
-            QaReviewResponse(
-                dataId = args.dataId,
-                companyId = args.companyId,
-                companyName = args.companyName,
-                framework = framework,
-                reportingPeriod = args.reportingPeriod,
-                timestamp = args.timestamp,
-                qaStatus = QaStatus.Pending,
-                qaJudgeUserId = args.qaJudgeUserId.toString(),
-                qaJudgeUserName = args.qaJudgeUserName,
-                datasetReviewId = args.judgementId.toString(),
-                numberQaReports = args.numberQaReports,
-                comment = args.comment,
-                triggeringUserId = null,
-                priorityOfAssociatedDataSourcing = args.priority,
-            )
-
-        val dataSourcingPriority =
-            DataSourcingPriorityByDataDimensions(
-                dataType = framework,
-                reportingPeriod = args.reportingPeriod,
-                companyId = args.companyId,
-                priority = args.priority,
-            )
-
-        return TestSetResult(qaReview, datasetJudgement, response, dataSourcingPriority)
-    }
-
+    
     private val testSet1 =
         buildTestSet(
             BuildTestSetArgs(
